pyproject.toml
[tool.poetry]
name = "eprel_collector"
version = "0.9.0"
description = "Collects information from European Product Registry for Energy Labelling"
authors = ["Oleg Isaev <oleg.isayev@gmail.com>"]
readme = "README.md"
repository = "https://github.com/oisaev/eprel_collector"

[tool.poetry.dependencies]
python = "~3.11"
aiohttp = "^3.9.1"
sqlalchemy = "^2.0.23"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
prettytable = "^3.9.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
isort = "^5.13.0"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_local_folder = ["core", "models", "utils"]
lines_after_imports = 2
line_length = 79
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true

[tool.black]
line-length = 79
skip-string-normalization = true
